+++ Code/client/Games/Skyrim/Forms/AlchemyItem.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "TESForm.h"
+
+struct AlchemyItem : TESForm
+{
+
+};